refactor: change file name format to kebab
